requirements.txt
coreapi==2.3.3
Django==4.2.5
djangorestframework==3.14.0
httpcore==1.0.2
httpx==0.26.0
requests==2.31.0